Refactor and rename FAT
