+++ dev/io.openliberty.checkpoint_fat_security_oidc_social/.settings/bndtools.core.prefs
