+++ dev/io.openliberty.checkpoint_fat_security_oidc_social/build.gradle
