+++ dev/io.openliberty.checkpoint_fat_security_oidc_social/publish/files/internalFeatures/securitylibertyinternals-1.0.mf
