+++ dev/io.openliberty.checkpoint_fat_security_oidc_social/publish/securitykeys/rsa_trust.jks
