+++ dev/io.openliberty.checkpoint_fat_security_oidc_social/publish/servers/com.ibm.ws.security.social_fat.LibertyOP.social/configs/server_withJwtSsoFeature.xml
